refactor(scale): ScoringController和ScoringService DTO化
- ScoreCalculateRequest添加answers字段
- 新增ScoreInterpretRequest
- ScoringService.calculateScore改为接受ScoreCalculateRequest
- ScoringService.interpretScore改为接受ScoreInterpretRequest
- ExamServiceImpl调用ScoringService适配新接口

+++ ps-scale/src/main/java/com/iotsic/ps/scale/dto/ScoreCalculateResponse.java
@@ -2,6 +2,7 @@ package com.iotsic.ps.scale.dto;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.util.Map;
 
 /**
@@ -21,7 +22,7 @@ public class ScoreCalculateResponse {
     /**
      * 各维度得分
      */
-    private Map<String, Integer> dimensionScores;
+    private Map<String, BigDecimal> dimensionScores;
 
     /**
      * 得分等级
